moving parser to self file
